+++ src/app/core/services/generator.factory.ts
@@ -1,0 +1,7 @@
+import {GeneratorService} from "./generator";
+import {InjectionToken} from "@angular/core";
+
+export function GeneratorFactory (n: number): (service: GeneratorService) => string {
+  return (service: GeneratorService) => service.generate(n);
+}
+export const generatedString = new InjectionToken('generatedString');