+++ src/app/core/services/local-storage.service.ts
@@ -1,0 +1,11 @@
+export class LocalStorageService {
+  constructor() {}
+
+  setItemToLocalStorage(key: string, value: string): void {
+    window.localStorage.setItem(key, value);
+  }
+
+  getItemFromLocalStorage(key: string): string | null {
+    return window.localStorage.getItem(key);
+  }
+}